# DepthLimiter stops using ThreadStatic and uses a list instead