This PR removes `[ThreadStatic]` in `DepthLimiter` and replaces it with a simple unidirectional list of limiters, ended by a null object (doing nothing). 

The original idea was to check whether it's worth it to replace it and I found some potential improvements. At the same time it looks like it leaks some allocations, but I cannot find it though 😕 . Anyway, I'm sharing `AllocationBenchmarks` below. Still, it requires an opinionated review and potentially running of some other benchmarks.

```ini
BenchmarkDotNet=v0.11.5, OS=Windows 10.0.17134.829 (1803/April2018Update/Redstone4)
Intel Core i7-6700HQ CPU 2.60GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
Frequency=2531247 Hz, Resolution=395.0622 ns, Timer=TSC
.NET Core SDK=2.2.300
  [Host]     : .NET Core 2.2.5 (CoreCLR 4.6.27617.05, CoreFX 4.6.27618.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.2.5 (CoreCLR 4.6.27617.05, CoreFX 4.6.27618.01), 64bit RyuJIT
```

#### Before

|               Method |        Mean |       Error |      StdDev |  Gen 0 | Allocated |
|--------------------- |------------:|------------:|------------:|-------:|----------:|
|             LogEmpty |    10.90 ns |   0.2576 ns |   0.3934 ns |      - |         - |
| LogEmptyWithEnricher |    73.15 ns |   1.3302 ns |   1.1792 ns | 0.0178 |      56 B |
|            LogScalar |   920.46 ns |  18.0878 ns |  20.8300 ns | 0.1192 |     376 B |
|        LogDictionary | 4,767.65 ns |  95.0796 ns | 180.8989 ns | 0.7095 |    2240 B |
|          LogSequence | 2,222.24 ns |  44.3338 ns | 121.3631 ns | 0.2594 |     824 B |
|         LogAnonymous | 9,579.28 ns | 188.7198 ns | 406.2393 ns | 1.0834 |    3440 B |

#### After

|               Method |         Mean |       Error |      StdDev |  Gen 0 | Allocated |
|--------------------- |-------------:|------------:|------------:|-------:|----------:|
|             LogEmpty |     9.626 ns |   0.2251 ns |   0.3699 ns |      - |         - |
| LogEmptyWithEnricher |    69.689 ns |   1.4173 ns |   1.4554 ns | 0.0178 |      56 B |
|            LogScalar |   827.965 ns |  29.0880 ns |  34.6273 ns | 0.1192 |     376 B |
|        LogDictionary | 4,332.959 ns |  70.4718 ns |  65.9194 ns | 0.7248 |    2288 B |
|          LogSequence | 1,952.709 ns |  41.2889 ns |  38.6217 ns | 0.2632 |     840 B |
|         LogAnonymous | 8,187.965 ns | 136.0692 ns | 127.2793 ns | 1.1139 |    3504 B |