@Scooletz 2019-06-27
I'm not happy with the benchmarks' results. They show some skews on my machine (depending on various factors). Anyway, will investigate allocs more.