@Scooletz 2019-06-28
Let's leave it with `[ThreadStatic]` for now. I can't find these allocs.